backend/requirements.txt
# framework
fastapi-slim==0.115.6

# database
aiosqlite==0.20.0
alembic==1.14.0
SQLAlchemy==2.0.36

# pydantic
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2

# server
uvicorn==0.34.0

# authentication
PyJWT==2.10.1
cryptography==44.0.0

# other
python-multipart==0.0.20
greenlet==3.1.1
orjson==3.10.14